feat: implement direct GBT solo mining architecture
- Replace buffer-heavy local_work with streaming FSM parser (gbt_fsm_parser)
- Add BIP22/BIP141/BIP341 coinbase builder with address validation (coinbase_builder)
- Integrate GBT client and failover handling into stratum_task
- Add unit test suites for FSM parser and coinbase generation
- Integrate GBT pool configuration (payoutAddress, minerTag) in OpenAPI and Axe-OS
- Revert unrelated drive-by changes to match master

+++ main/task_monitor.c
@@ -48,8 +48,7 @@ void task_monitor_task(void *pvParameters) {
 
         // Warn if task count changed (possible create/destroy)
         if (num_tasks2 != num_tasks1) {
-            ESP_LOGW(TAG, "Task count changed (%lu -> %lu); stats may be incomplete for destroyed tasks",
-                     (unsigned long)num_tasks1, (unsigned long)num_tasks2);
+            ESP_LOGW(TAG, "Task count changed (%u -> %u); stats may be incomplete for destroyed tasks", num_tasks1, num_tasks2);
         }
 
         // Print header
